unsigned long -> unsigned int
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@3649 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ Binaries/dcmtk-source/config/cfunix.h
@@ -355,10 +355,10 @@ typedef unsigned char uchar;
 typedef unsigned int uint;
 #endif
 
-/* Define `ulong' to `unsigned long' if <sys/types.h> does not define. */
+/* Define `ulong' to `unsigned int' if <sys/types.h> does not define. */
 //#define HAVE_NO_TYPEDEF_ULONG 1
 #ifdef HAVE_NO_TYPEDEF_ULONG
-typedef unsigned long ulong;
+typedef unsigned int ulong;
 #endif
 
 /* Define `ushort' to `unsigned short' if <sys/types.h> does not define. */
